- Subscriptions now has a menu option for refreshing the subscriptions, and will automatically refresh the subscriptions when they are added and edited - Subscriptions can now be edited - Add/Edit Subscription dialog now supports vertical scrolling for landscape mode and small screens - Host address field will no longer provide auto-complete suggestions

+++ GrowlForAndroid/src/com/growlforandroid/client/Subscriptions.java
@@ -5,32 +5,29 @@ import com.growlforandroid.common.GrowlApplication;
 import com.growlforandroid.common.GrowlNotification;
 import com.growlforandroid.common.NotificationType;
 
-import android.app.AlertDialog;
-import android.app.Dialog;
-import android.app.ListActivity;
-import android.content.DialogInterface;
+import android.app.*;
+import android.app.AlertDialog.Builder;
+import android.content.*;
 import android.database.Cursor;
 import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.Menu;
-import android.view.MenuItem;
-import android.view.View;
-import android.widget.EditText;
-import android.widget.ListView;
-import android.widget.SimpleCursorAdapter;
+import android.util.Log;
+import android.view.*;
+import android.widget.*;
 
 public class Subscriptions
 	extends ListActivity
 	implements GrowlListenerService.StatusChangedHandler {
 	
 	private final int DIALOG_ADD_SUBSCRIPTION = 0;
-	private final int DIALOG_ITEM_MENU = 1;
+	private final int DIALOG_EDIT_SUBSCRIPTION = 1;
+	private final int DIALOG_ITEM_MENU = 2;
 	
 	private ListenerServiceConnection _service;
 	
 	private Database _database;
 	private Cursor _cursor;
 	private MenuItem _mniAddSubscription;
+	private MenuItem _mniSubscribeNow;
 	private long _itemId = -1;
 	
     @Override
@@ -106,6 +103,8 @@ public class Subscriptions
     public boolean onCreateOptionsMenu(Menu menu) {
     	_mniAddSubscription = menu.add(R.string.subscriptions_add_subscription);
     	_mniAddSubscription.setIcon(android.R.drawable.ic_menu_add);
+    	_mniSubscribeNow = menu.add(R.string.subscriptions_subscribe_now);
+    	_mniSubscribeNow.setIcon(android.R.drawable.ic_menu_rotate);
     	return true;
     }
     
@@ -114,6 +113,9 @@ public class Subscriptions
     	if (item == _mniAddSubscription) {
     		showDialog(DIALOG_ADD_SUBSCRIPTION);
     		return true;
+    	} else if (item == _mniSubscribeNow) {
+    		if (_service.isRunning())
+    			_service.subscribeNow();
     	}
     	return false;
     }
@@ -128,44 +130,36 @@ public class Subscriptions
     public Dialog onCreateDialog(int id) {
     	switch (id) {
 	    	case DIALOG_ADD_SUBSCRIPTION:
-	        	LayoutInflater factory = LayoutInflater.from(this);
-	            final View textEntryView = factory.inflate(R.layout.add_subscription_dialog, null);
-	            final EditText txtName = (EditText) textEntryView.findViewById(R.id.txtSubscriptionName);
-	            final EditText txtAddress = (EditText) textEntryView.findViewById(R.id.txtSubscriptionAddress);
-	            final EditText txtPassword = (EditText) textEntryView.findViewById(R.id.txtSubscriptionPassword);
-	            
-	            return new AlertDialog.Builder(this)
-	                .setTitle(R.string.add_subscription_dialog_title)
-	                .setView(textEntryView)
-	                .setPositiveButton(R.string.dialog_add, new DialogInterface.OnClickListener() {
-	                    public void onClick(DialogInterface dialog, int whichButton) {
-	                    	// Add the password to the database
-	                    	String name = txtName.getText().toString();
-	                    	String address = txtAddress.getText().toString();
-	                    	String password = txtPassword.getText().toString();
-	                    	String status = getText(R.string.subscriptions_status_unregistered).toString();
-	                    	if (name.equals("") || address.equals("") || password.equals("")) {
-	                    		return;
-	                    	}	                    	
-	                    	_database.insertSubscription(name, address, password, status);
-	                    	refresh();
-	                    }
-	                })
-	                .setNegativeButton(R.string.dialog_cancel, new DialogInterface.OnClickListener() {
-	                    public void onClick(DialogInterface dialog, int whichButton) {
-	                        /* User clicked cancel so don't do anything */
-	                    }
-	                })
-	                .create();
-	            
+	    		return new EditDialog(this).create();
+	        
+	    	case DIALOG_EDIT_SUBSCRIPTION:
+	    		Cursor subscription = _database.getSubscription(_itemId);
+	    		if (subscription.moveToFirst()) {
+		    		String name = subscription.getString(subscription.getColumnIndex(Database.KEY_NAME));
+		        	String address = subscription.getString(subscription.getColumnIndex(Database.KEY_ADDRESS));
+		        	String password = subscription.getString(subscription.getColumnIndex(Database.KEY_PASSWORD));
+		    		subscription.close();
+		        	return new EditDialog(this, _itemId, name, address, password).create();
+	    		}
+	    		subscription.close();
+	    		break;
+		    	
+	    		
 	    	case DIALOG_ITEM_MENU:
 	            return new AlertDialog.Builder(this)
 	                .setTitle(R.string.subscriptions_menu_title)
 	                .setItems(R.array.subscriptions_menu, new DialogInterface.OnClickListener() {
 	                    public void onClick(DialogInterface dialog, int which) {
-	                    	// Only option is delete
-	                    	_database.deleteSubscription(_itemId);
-	                    	refresh();
+	                    	if (which == 0) {
+	                    		// Edit
+	                    		showDialog(DIALOG_EDIT_SUBSCRIPTION);
+	                    	} else if (which == 1) {
+	                    		// Delete
+	                    		_database.deleteSubscription(_itemId);
+		                    	refresh();	
+	                    	} else {
+	                    		Log.w("Subscriptions.ItemMenu.onClick", "Unknown option " + which);
+	                    	}
 	                    }
 	                })
 	                .create();
@@ -173,6 +167,98 @@ public class Subscriptions
     	
     	return null;
     }
+
+	private class EditDialog {
+		private final Context _context;
+		private LayoutInflater _factory;
+		private View _textEntryView;
+		private EditText _txtName;
+		private EditText _txtAddress;
+		private EditText _txtPassword;
+		
+		private Builder _builder;
+
+		private long _id;
+		
+		public EditDialog(Context context) {
+			_context = context;
+			initialise();
+			
+			_builder.setTitle(R.string.subscriptions_add_dialog_title);
+		    _builder.setPositiveButton(R.string.dialog_add, new DialogInterface.OnClickListener() {
+		        public void onClick(DialogInterface dialog, int whichButton) {
+		        	// Get the field values from the dialog
+		        	String name = _txtName.getText().toString();
+		        	String address = _txtAddress.getText().toString();
+		        	String password = _txtPassword.getText().toString();
+		        	String status = getText(R.string.subscriptions_status_unregistered).toString();
+		        	if (name.equals("") || address.equals("") || password.equals("")) {
+		        		return;
+		        	}	                    	
+		        	
+		        	// Add the subscription to the database, then refresh the list
+		        	_database.insertSubscription(name, address, password, status);
+		        	refresh();
+		        	
+		        	// Get the service to subscribe now
+		        	if (_service.isRunning())
+		        		_service.subscribeNow();
+		        }
+		    });
+		}
+		
+		public EditDialog(Context context, long id, String name, String address, String password) {
+			_context = context;
+			_id = id;
+			initialise();
+			
+			_txtName.setText(name);
+			_txtAddress.setText(address);
+			_txtPassword.setText(password);
+			
+			_builder.setTitle(R.string.subscriptions_edit_dialog_title);
+		    _builder.setPositiveButton(R.string.dialog_save, new DialogInterface.OnClickListener() {
+		        public void onClick(DialogInterface dialog, int whichButton) {
+		        	// Get the field values from the dialog
+		        	String name = _txtName.getText().toString();
+		        	String address = _txtAddress.getText().toString();
+		        	String password = _txtPassword.getText().toString();
+		        	String status = getText(R.string.subscriptions_status_unregistered).toString();
+		        	if (name.equals("") || address.equals("") || password.equals("")) {
+		        		return;
+		        	}	                    	
+		        	
+		        	// Add the subscription to the database, then refresh the list
+		        	_database.updateSubscription(_id, name, address, password, status);
+		        	refresh();
+		        	
+		        	// Get the service to subscribe now
+		        	if (_service.isRunning())
+		        		_service.subscribeNow();
+		        }
+		    });
+		}
+		
+		private void initialise() {
+			_factory = LayoutInflater.from(_context);
+			_textEntryView = _factory.inflate(R.layout.add_subscription_dialog, null);
+			_txtName = (EditText) _textEntryView.findViewById(R.id.txtSubscriptionName);
+			_txtAddress = (EditText) _textEntryView.findViewById(R.id.txtSubscriptionAddress);
+			_txtPassword = (EditText) _textEntryView.findViewById(R.id.txtSubscriptionPassword);
+			
+			_builder = new AlertDialog.Builder(_context)			    
+			    .setView(_textEntryView)
+			    .setNegativeButton(R.string.dialog_cancel, new DialogInterface.OnClickListener() {
+			        public void onClick(DialogInterface dialog, int whichButton) {
+			            /* User clicked cancel so don't do anything */
+			        }
+			    });
+		}
+		
+		public Dialog create() {
+			return _builder.create();
+		}
+	}
     
 
     public void onApplicationRegistered(GrowlApplication app) {
